+++ R/TuningInstanceMulticrit.R
@@ -21,6 +21,7 @@
 #' @template param_allow_hotstart
 #' @template param_keep_hotstart_stack
 #' @template param_evaluate_default
+#' @template param_callbacks
 #' @template param_xdt
 #' @template param_learner_param_vals
 #'
@@ -52,7 +53,7 @@ TuningInstanceMultiCrit = R6Class("TuningInstanceMultiCrit",
 
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
-    initialize = function(task, learner, resampling, measures, terminator, search_space = NULL, store_benchmark_result = TRUE, store_models = FALSE, check_values = FALSE, allow_hotstart = FALSE, keep_hotstart_stack = FALSE, evaluate_default = FALSE) {
+    initialize = function(task, learner, resampling, measures, terminator, search_space = NULL, store_benchmark_result = TRUE, store_models = FALSE, check_values = FALSE, allow_hotstart = FALSE, keep_hotstart_stack = FALSE, evaluate_default = FALSE, callbacks = list()) {
       private$.evaluate_default = assert_flag(evaluate_default)
       learner = assert_learner(as_learner(learner, clone = TRUE))
 
@@ -72,9 +73,9 @@ TuningInstanceMultiCrit = R6Class("TuningInstanceMultiCrit",
 
       # initialized specialized tuning archive and objective
       archive = ArchiveTuning$new(search_space, codomain, check_values)
-      objective = ObjectiveTuning$new(task, learner, resampling, measures, store_benchmark_result, store_models, check_values, allow_hotstart, keep_hotstart_stack, archive)
+      objective = ObjectiveTuning$new(task, learner, resampling, measures, store_benchmark_result, store_models, check_values, allow_hotstart, keep_hotstart_stack, archive, callbacks)
 
-      super$initialize(objective, search_space, terminator)
+      super$initialize(objective, search_space, terminator, callbacks = callbacks)
       # super class of instance initializes default archive, overwrite with tuning archive
       self$archive = archive
     },