+++ R/tune.R
@@ -45,6 +45,7 @@
 #' @template param_allow_hotstart
 #' @template param_keep_hotstart_stack
 #' @template param_evaluate_default
+#' @template param_callbacks
 #'
 #' @export
 #' @examples
@@ -63,7 +64,7 @@
 #'
 #' # get result
 #' instance$result
-tune = function(method, task, learner, resampling, measures = NULL, term_evals = NULL, term_time = NULL, terminator = NULL, search_space = NULL, store_benchmark_result = TRUE, store_models = FALSE, check_values = FALSE, allow_hotstart = FALSE, keep_hotstart_stack = FALSE, evaluate_default = FALSE, ...) {
+tune = function(method, task, learner, resampling, measures = NULL, term_evals = NULL, term_time = NULL, terminator = NULL, search_space = NULL, store_benchmark_result = TRUE, store_models = FALSE, check_values = FALSE, allow_hotstart = FALSE, keep_hotstart_stack = FALSE, evaluate_default = FALSE, callbacks = list(), ...) {
   tuner = if (is.character(method)) {
     assert_choice(method, mlr_tuners$keys())
     tnr(method, ...)
@@ -73,7 +74,7 @@ tune = function(method, task, learner, resampling, measures = NULL, term_evals =
   terminator = terminator %??% terminator_selection(term_evals, term_time)
 
   TuningInstance = if (!is.list(measures)) TuningInstanceSingleCrit else TuningInstanceMultiCrit
-  instance = TuningInstance$new(task, learner, resampling, measures, terminator, search_space, store_benchmark_result, store_models, check_values, allow_hotstart, keep_hotstart_stack, evaluate_default)
+  instance = TuningInstance$new(task, learner, resampling, measures, terminator, search_space, store_benchmark_result, store_models, check_values, allow_hotstart, keep_hotstart_stack, evaluate_default, callbacks)
 
   tuner$optimize(instance)
   instance
@@ -93,4 +94,3 @@ terminator_selection = function(term_evals, term_time) {
     trm("run_time", secs = term_time)
   }
 }
-