+++ man-roxygen/param_callbacks.R
@@ -1,0 +1,2 @@
+#' @param callbacks (list of [CallbackTuning])\cr
+#'   List of callbacks.